api/requirements.txt
cffi==1.14.5
click==8.0.0
colorama==0.4.4
cryptography==3.4.7
dnspython==2.1.0
Flask==2.0.0
Flask-Cors==3.0.10
itsdangerous==2.0.1
Jinja2==3.0.0
MarkupSafe==2.0.0
pycparser==2.20
PyJWT==2.1.0
pymongo==3.11.4
six==1.16.0
Werkzeug==2.0.1
